requirements.txt
httpx>=0.27
playwright>=1.40
anthropic>=0.40
openai>=1.50
apscheduler>=3.10,<4.0
aiosqlite>=0.20
pyyaml>=6.0
pydantic>=2.5
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=0.23
